src/machine learning/requirements.txt
filetype
python-dotenv
flask